The on_.*_match decorators can now be decorated too.
This allows functions to be called on more than one message type,
or on matching one of many regular expressions.

It looks like this:

        @on_channel_match("foo")
        @on_addressed_match("bar")
        def func(...):
		...

func will be called upon finding 'foo' in a channel message, or 'bar'
in an addressed message.

The function will only be called once per message, even if it is matched
more than once.